+++ test_sec_filings/sec_filings/test_section_api.py
@@ -1,5 +1,7 @@
 import os
 import pytest
+import csv
+from io import StringIO
 
 from fastapi.testclient import TestClient
 
@@ -226,6 +228,41 @@ def test_section_narrative_api_with_headers(
     )
 
     assert response.status_code == response_status
+
+
+@pytest.mark.parametrize(
+    "form_type, response_type, section",
+    [
+        ("10-K", "text/csv", "RISK_FACTORS"),
+        ("10-Q", "text/csv", "RISK_FACTORS"),
+        ("S-1", "text/csv", "RISK_FACTORS"),
+        ("10-K", "text/csv", "_ALL"),
+        ("10-Q", "text/csv", "_ALL"),
+        ("S-1", "text/csv", "_ALL"),
+    ],
+)
+def test_section_narrative_api_with_text_csv_response_type(form_type, response_type, section, tmpdir):
+    sample_document = generate_sample_document(form_type)
+    filename = os.path.join(tmpdir.dirname, "wilderness.xbrl")
+    with open(filename, "w") as f:
+        f.write(sample_document)
+
+    # NOTE(robinson) - Reset the rate limit to avoid 429s in tests
+    app.state.limiter.reset()
+    client = TestClient(app)
+    response = client.post(
+        SECTION_ROUTE,
+        files=[("text_files", (filename, open(filename, "rb"), "text/plain"))],
+        data={"output_format": response_type, "section": [section]}
+    )
+    assert response.status_code == 200
+
+    response_csv = csv.DictReader(StringIO(response.json()), delimiter=',')
+    response_list = list(response_csv)
+    
+    assert [x['section'] for x in response_list]
+    assert [x['element_type'] for x in response_list]
+    assert [x['text'] for x in response_list]
 
 
 def test_section_narrative_api_health_check():