+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.2-dev
+
+* Update `pipeline_api` to support `text/csv` response
+* Add new function `convert_to_isd_csv`
+* Add test function `test_section_narrative_api_with_text_csv_response_type`
+
 ## 0.2.1-dev
 
 * More general (non-pipeline-specific) way of starting the app